Cherry Pick Training Mode Bug Fixes From Master (#1273)
Bring over these bug fixes from master: 1. Add the ability for database modules to initialize their shadow table configurations. This is mostly useful for writing tests. 2.  Throw an exception if the ClientContext is null during database initialization. 3. Add support for shadow tables to create tag and/or augmented columns when they are created.

Co-authored-by: Joe Brownlee <Joe.Brownlee@petco.com>