PDPOS-6552: Fix scanning bug related to focus retention in search-expand-input component (#1644)
For the following scenario:
- User is on Blind Return screen
- User hand keys a sku value into the scan-expand-input component
- User presses return to accept the sku
- A dialog shows asking the user if they want to narrow the transaction associated with the sku by loyalty number, or by scanning the receipt, or a third option.
- User chooses to scan a receipt, which takes them back them back to the blind Return screen.
- At this point the focus comes back to the input field on the search-expand-input, but there is no visual indication of that because the input field is contracted.
- User scans a receipt, but the data from the scan is read in reverse order of the barcode.  I am not exactly sure why that happens, but if the scan field did not have focus (and, in my opinion, it shouldn't since the primary 'Checkout' button on that screen is requesting focus), the ScannerService would handle the scan in the same fashion it would normally handle it when the search-expand-input field does not have focus.
- Since the scan digits are read in reverse order, the transaction is not found and the user gets an error.

While troubleshooting this I noticed that the primary button on the page, which is the Checkout button, was requesting focus via cdkFocusInitial Directive, but this error was showing in the log: `Element matching '[cdkFocusInitial]' is not focusable. HTMLButtonElement`. I looked at which component was responsible for that error, which was the primary-button component.  I changed the cdkFocusInitial directive on the the button element to `autofocus` as per HTML guidelines.  In my testing this seems to indirectly fix the problem because now when the blind return screen is shown after the dialog is dismissed the Checkout button is getting the focus and allowing the subsequent scan to be handled by the ScanService.

- Also fixing an exception being generated in the ActionSyncMessagePropertyStrategy: `java.lang.ClassCastException: org.jumpmind.pos.core.ui.Toast cannot be cast to org.jumpmind.pos.core.ui.UIMessage`